Platform-specific fixes
