+++ streams/shared/src/main/scala/zio/stream/experimental/ZStream.scala
@@ -1,11 +1,8 @@
 package zio.stream.experimental
 
-import java.io.IOException
-import java.io.InputStream
 import java.{ util => ju }
 
 import zio._
-import zio.blocking.Blocking
 import zio.internal.UniqueKey
 import zio.stm.TQueue
 
@@ -1833,7 +1830,7 @@ abstract class ZStream[-R, +E, +O](
   }
 }
 
-object ZStream {
+object ZStream extends ZStreamPlatformSpecificConstructors {
 
   /**
    * The default chunk size used by the various combinators and constructors of [[ZStreamChunk]].
@@ -2066,44 +2063,6 @@ object ZStream {
           if (_) Pull.end              -> true
           else fa.map(Chunk.single(_)) -> true
         }.flatten
-      } yield pull
-    }
-
-  /**
-   * Creates a stream from a [[java.io.InputStream]]
-   */
-  def fromInputStream(
-    is: => InputStream,
-    chunkSize: Int = ZStream.DefaultChunkSize
-  ): ZStream[Blocking, IOException, Byte] =
-    ZStream {
-      for {
-        done       <- Ref.make(false).toManaged_
-        buf        <- Ref.make(Array.ofDim[Byte](chunkSize)).toManaged_
-        capturedIs <- Managed.effectTotal(is)
-        pull = {
-          def go: ZIO[Blocking, Option[IOException], Chunk[Byte]] = done.get.flatMap {
-            if (_) Pull.end
-            else
-              for {
-                bufArray <- buf.get
-                bytesRead <- blocking
-                              .effectBlockingInterrupt(capturedIs.read(bufArray))
-                              .refineToOrDie[IOException]
-                              .mapError(Some(_))
-                bytes <- if (bytesRead < 0)
-                          done.set(true) *> Pull.end
-                        else if (bytesRead == 0)
-                          go
-                        else if (bytesRead < bufArray.length)
-                          Pull.emit(Chunk.fromArray(bufArray).take(bytesRead))
-                        else
-                          Pull.emit(Chunk.fromArray(bufArray))
-              } yield bytes
-          }
-
-          go
-        }
       } yield pull
     }
